Criado JavaScript modules
Iniciado a conversão do JavaScript para modules.
