+++ js/scripts.js
@@ -1,0 +1,7 @@
+import initBurgerMenu from './modules/menuHamburger.js';
+import initTimeCheck from './modules/timeChecker.js';
+import initSmoothScroll from './modules/smoothScroll.js';
+
+initBurgerMenu();
+initTimeCheck();
+initSmoothScroll();
