+++ matricula.html
@@ -26,8 +26,8 @@
     <link href="https://fonts.googleapis.com/css2?family=Alice&family=Nunito:wght@400;700&display=swap" rel="stylesheet">
     <link rel="shortcut icon" href="./favicon.svg" type="image/svg+xml">
     <link rel="stylesheet" href="./css/style.min.css">
+    <script defer type="module" src="./js/scripts.js"></script>
     <script defer src="./js/form.js"></script>
-    <script defer src="./js/script.js"></script>
     <script>document.documentElement.classList.add('js')</script>
 </head>
